色々消した
