+++ frontend/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "../components/playMusic";
+import App from "../components/search";
 import "../styles/index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
