--- frontend/styles/design.css
@@ -1,77 +1,0 @@
-.App {
-    text-align: center;
-  }
-  
-  .App-logo {
-    height: 40vmin;
-    pointer-events: none;
-  }
-  
-  @media (prefers-reduced-motion: no-preference) {
-    .App-logo {
-      animation: App-logo-spin infinite 20s linear;
-    }
-  }
-  
-  .App-header {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-  }
-  /*
-  .App-link {
-    color: #61dafb;
-  }
-  
-  @keyframes App-logo-spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  } */
-  #btn{
-    background-color: #61dafb;
-  }
-  #test{
-    display: inline;
-  }
-  .bg_pattern {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background-color: #ffe9a7;
-    opacity: 0.4;
-    z-index: -1;
-  }
-  .Diagonal {
-    background: repeating-linear-gradient( 45deg, #ffc107, #ffc107 10px, #ffe9a7 10px, #ffe9a7 50px );
-  }
-  #resultArea{
-    color: #000;
-  }
-  #fileList{
-    color: #000;
-  }
-  #name{
-    width: 30vw;
-    /* background-image: url(CD.png); */
-  }
-  #local{
-    width: 30vw;
-  }
-  #weather{
-    width: 30vw;
-  }
-  #research{
-    width: 80vw;
-    background-color: aliceblue;
-    align-items: center;
-    top: 0;
-  }
